# CLN/PERF: group quantile