--- a/pandas/_libs/groupby.pyi
+++ b/pandas/_libs/groupby.pyi
@@ -2,6 +2,8 @@ from typing import Literal
 
 import numpy as np
 
+from pandas._typing import npt
+
 def group_median_float64(
     out: np.ndarray,  # ndarray[float64_t, ndim=2]
     counts: np.ndarray,  # ndarray[int64_t]
@@ -84,11 +86,12 @@ def group_ohlc(
     min_count: int = ...,
 ) -> None: ...
 def group_quantile(
-    out: np.ndarray,  # ndarray[float64_t, ndim=2]
+    out: npt.NDArray[np.float64],
     values: np.ndarray,  # ndarray[numeric, ndim=1]
-    labels: np.ndarray,  # ndarray[int64_t]
-    mask: np.ndarray,  # ndarray[uint8_t]
-    qs: np.ndarray,  # const float64_t[:]
+    labels: npt.NDArray[np.intp],
+    mask: npt.NDArray[np.uint8],
+    sort_arr: npt.NDArray[np.intp],  # const
+    qs: npt.NDArray[np.float64],  # const
     interpolation: Literal["linear", "lower", "higher", "nearest", "midpoint"],
 ) -> None: ...
 def group_last(