--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -774,6 +774,7 @@ def group_quantile(ndarray[float64_t, ndim=2] out,
                    ndarray[numeric, ndim=1] values,
                    ndarray[intp_t] labels,
                    ndarray[uint8_t] mask,
+                   const intp_t[:] sort_arr,
                    const float64_t[:] qs,
                    str interpolation) -> None:
     """
@@ -787,6 +788,8 @@ def group_quantile(ndarray[float64_t, ndim=2] out,
         Array containing the values to apply the function against.
     labels : ndarray[np.intp]
         Array containing the unique group labels.
+    sort_arr : ndarray[np.intp]
+        Indices describing sort order by values and labels.
     qs : ndarray[float64_t]
         The quantile values to search for.
     interpolation : {'linear', 'lower', 'highest', 'nearest', 'midpoint'}
@@ -800,9 +803,9 @@ def group_quantile(ndarray[float64_t, ndim=2] out,
         Py_ssize_t i, N=len(labels), ngroups, grp_sz, non_na_sz, k, nqs
         Py_ssize_t grp_start=0, idx=0
         intp_t lab
-        uint8_t interp
+        InterpolationEnumType interp
         float64_t q_val, q_idx, frac, val, next_val
-        ndarray[int64_t] counts, non_na_counts, sort_arr
+        int64_t[::1] counts, non_na_counts
 
     assert values.shape[0] == N
 
@@ -836,16 +839,6 @@ def group_quantile(ndarray[float64_t, ndim=2] out,
             counts[lab] += 1
             if not mask[i]:
                 non_na_counts[lab] += 1
-
-    # Get an index of values sorted by labels and then values
-    if labels.any():
-        # Put '-1' (NaN) labels as the last group so it does not interfere
-        # with the calculations.
-        labels_for_lexsort = np.where(labels == -1, labels.max() + 1, labels)
-    else:
-        labels_for_lexsort = labels
-    order = (values, labels_for_lexsort)
-    sort_arr = np.lexsort(order).astype(np.int64, copy=False)
 
     with nogil:
         for i in range(ngroups):