… pandas/_libs/groupby.html